requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
pymysql==1.1.1
mysql-connector-python==9.2.0
mysqlclient==2.2.7
whitenoise==6.6.0
nose==1.3.7
Pillow==11.1.0
django-nose==1.4.7
coverage==7.5.1
pynose==1.5.1
cryptography
django-filter
factory-boy
